Renamed carcasses to carcasses_and_poo
